+++ R/VERBS-where.R
@@ -16,8 +16,10 @@ where <- function(.data, ...) {
 #' @rdname where-table.express
 #' @export
 #' @importFrom rlang enquo
+#' @importFrom rlang is_missing
 #' @importFrom rlang quo_get_expr
 #'
+#' @param which Passed to [data.table::data.table].
 #' @param .collapse A boolean function which will be used to "concatenate" all conditions in `...`.
 #' @template parse-arg
 #' @template chain-arg
@@ -32,7 +34,7 @@ where <- function(.data, ...) {
 #'     start_expr %>%
 #'     where(vs == 0, am == 1)
 #'
-where.ExprBuilder <- function(.data, ..., .collapse = `&`,
+where.ExprBuilder <- function(.data, ..., which = FALSE, .collapse = `&`,
                               .parse = getOption("table.express.parse", FALSE),
                               .chain = getOption("table.express.chain", TRUE))
 {
@@ -52,5 +54,10 @@ where.ExprBuilder <- function(.data, ..., .collapse = `&`,
         clause <- reduce_expr(clause[-1L], first_where, .collapse, .parse = .parse)
     }
 
-    .data$set_where(clause, .chain)
+    .data <- .data$set_where(clause, .chain)
+    if (!rlang::is_missing(which)) {
+        frame_append(.data, which = !!which, .parse = FALSE)
+    }
+
+    .data
 }